Link to react-intersection-observer
Thanks for the hook example! It was the original inspiration for upgrading [https://github.com/thebuilder/react-intersection-observer](https://github.com/thebuilder/react-intersection-observer) to support hooks. 🙌 

The example is a really nice starting point, but also limited if you were to use it in production. `react-intersection-observer` supports all configuration options for IntersectionObserver, and the hook correctly handles creating new observers if props/ref changes.

I think it would be a nice addition to provide a link for users looking to use an IntersectionObserver.